osobny skrypt do player health
