+++ Assets/Scripts/PlayerMovement.cs
@@ -3,10 +3,6 @@ using UnityEngine.AI;
 
 public class PlayerMovement : MonoBehaviour
 {
-    public int maxHealth = 20;
-    public int currentHealth;
-
-    public healthBar healthbar;
     enum PlayerAnimation { Idle, Run }
 
     NavMeshAgent agent;
@@ -18,31 +14,30 @@ public class PlayerMovement : MonoBehaviour
     [SerializeField] LayerMask clickableLayers;
 
     float lookRotationSpeed = 8f;
-    
-    void Awake() 
+
+    void Awake()
     {
         InputManager.Instance.GetInput().Main.Move.performed += input => ClickToMove();
     }
 
-    private void Start() {
+    private void Start()
+    {
         agent = GetComponent<NavMeshAgent>();
         animator = GetComponent<Animator>();
-        currentHealth = maxHealth;
-        healthbar.SetMaxHealth(maxHealth);
     }
 
     void ClickToMove()
     {
         RaycastHit hit;
-        if (Physics.Raycast(Camera.main.ScreenPointToRay(Input.mousePosition), out hit, 100, clickableLayers)) 
+        if (Physics.Raycast(Camera.main.ScreenPointToRay(Input.mousePosition), out hit, 100, clickableLayers))
         {
             agent.destination = hit.point;
-            if(clickEffect != null)
-                Instantiate(clickEffect, hit.point + new Vector3(0, 0.1f, 0), clickEffect.transform.rotation); 
+            if (clickEffect != null)
+                Instantiate(clickEffect, hit.point + new Vector3(0, 0.1f, 0), clickEffect.transform.rotation);
         }
     }
 
-    void Update() 
+    void Update()
     {
         if (destination != agent.destination)
         {
@@ -50,32 +45,21 @@ public class PlayerMovement : MonoBehaviour
             FaceTarget();
         }
         SetAnimations();
-        if (Input.GetKeyDown(KeyCode.Space))
-        {
-            TakeDamage(5);
-        }
     }
 
     void FaceTarget()
-	{
+    {
         Vector3 direction = (agent.destination - transform.position).normalized;
         Quaternion lookRotation = Quaternion.LookRotation(new Vector3(direction.x, 0, direction.z));
         transform.rotation = lookRotation;
-	}
+    }
 
     void SetAnimations()
     {
-        if(agent.velocity == Vector3.zero)
+        if (agent.velocity == Vector3.zero)
             animator.Play(PlayerAnimation.Idle.ToString());
         else
             animator.Play(PlayerAnimation.Run.ToString());
-    }
-
-    void TakeDamage(int damage)
-    {
-        currentHealth -= damage;
-
-        healthbar.SetHealth(currentHealth);
     }
 
     // private void OnDestroy()
