+++ Assets/Scripts/player_health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9bee0b4c4f3c57c4a97906a810537bba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
